1. Return additional token address and crowdsale address to frontend 2. Add flags to control deployment.

+++ script/sqlite.js
@@ -1,5 +1,6 @@
 var sqlite3=require('sqlite3-promise')
 var db = null;
+var global = require('./global');
 
 var SqliteHandler = {
   async loadDB(dbFilePath) {
@@ -48,16 +49,17 @@ var SqliteHandler = {
       addressListStr = row[0].data;
       rowId = row[0].id;
 
-      let queryDelete = 'DELETE FROM ' + addressTableName + ' WHERE ROWID=' + rowId;
-      try {
-        // TEST
-        await db.runAsync(queryDelete);
-        console.log('pop ' + addressTableName + ' OK - addressList:', addressListStr);
-      } catch (err) {
-        console.log('pop ' + addressTableName + ' Error: ', err);
-        return null;
+      if (global.DELETE_ADDRESS) {
+        let queryDelete = 'DELETE FROM ' + addressTableName + ' WHERE ROWID=' + rowId;
+        try {
+          // TEST
+          await db.runAsync(queryDelete);
+          console.log('pop ' + addressTableName + ' OK - addressList:', addressListStr);
+        } catch (err) {
+          console.log('pop ' + addressTableName + ' Error: ', err);
+          return null;
+        }
       }
-
     } catch (err) {
       console.log('pop ' + addressTableName + ' Error: ', err);
       return null;
